requirements.txt
xlrd
pandas
tensorflow-gpu
scipy
matplotlib
tqdm
sklearn
seaborn
